+++ src/Map.ts
@@ -8,6 +8,8 @@ import type {
   MapDataEvent,
   Tile,
   RasterDEMSourceSpecification,
+  TerrainSpecification,
+  MapTerrainEvent,
 } from "maplibre-gl";
 import { v4 as uuidv4 } from "uuid";
 import { ReferenceMapStyle, MapStyleVariant } from "@maptiler/client";
@@ -33,6 +35,15 @@ import { FullscreenControl } from "./FullscreenControl";
 function sleepAsync(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
+
+export type LoadWithTerrainEvent = {
+  type: "loadWithTerrain";
+  target: Map;
+  terrain: {
+    source: string;
+    exaggeration: number;
+  };
+};
 
 // StyleSwapOptions is not exported by Maplibre, but we can redefine it (used for setStyle)
 export type TransformStyleFunction = (
@@ -257,6 +268,11 @@ export class Map extends maplibregl.Map {
         console.warn(e.message);
       }
 
+      // A more precise localization
+
+      // This more advanced localization is commented out because the easeTo animation
+      // triggers an error if the terrain grow is enabled (due to being nable to project the center while moving)
+
       // Then, the get a more precise location, we rely on the browser location, but only if it was already granted
       // before (we don't want to ask wih a popup at launch time)
       const locationResult = await navigator.permissions.query({
@@ -272,11 +288,21 @@ export class Map extends maplibregl.Map {
               return;
             }
 
-            this.easeTo({
-              center: [data.coords.longitude, data.coords.latitude],
-              zoom: options.zoom || 12,
-              duration: 2000,
-            });
+            if (this.terrain) {
+              this.easeTo({
+                center: [data.coords.longitude, data.coords.latitude],
+                zoom: options.zoom || 12,
+                duration: 2000,
+              });
+            } else {
+              this.once("terrain", () => {
+                this.easeTo({
+                  center: [data.coords.longitude, data.coords.latitude],
+                  zoom: options.zoom || 12,
+                  duration: 2000,
+                });
+              });
+            }
           },
 
           // error callback
@@ -440,12 +466,83 @@ export class Map extends maplibregl.Map {
       }
     });
 
+    // Creating a custom event: "loadWithTerrain"
+    // that fires only once when both:
+    // - the map has full loaded (corresponds to the the "load" event)
+    // - the terrain has loaded (corresponds to the "terrain" event with terrain beion non-null)
+    // This custom event is necessary to wait for when the map is instanciated with `terrain: true`
+    // and some animation (flyTo, easeTo) are running from the begining.
+    let loadEventTriggered = false;
+    let terrainEventTriggered = false;
+    let terrainEventData: LoadWithTerrainEvent = null;
+
+    this.once("load", (_) => {
+      loadEventTriggered = true;
+      if (terrainEventTriggered) {
+        this.fire("loadWithTerrain", terrainEventData);
+      }
+    });
+
+    const terrainCallback = (evt) => {
+      if (!evt.terrain) return;
+      terrainEventTriggered = true;
+      terrainEventData = {
+        type: "loadWithTerrain",
+        target: this,
+        terrain: evt.terrain,
+      };
+      this.off("terrain", terrainCallback);
+
+      if (loadEventTriggered) {
+        this.fire("loadWithTerrain", terrainEventData as LoadWithTerrainEvent);
+      }
+    };
+
+    this.on("terrain", terrainCallback);
+
     // enable 3D terrain if provided in options
     if (options.terrain) {
       this.enableTerrain(
         options.terrainExaggeration ?? this.terrainExaggeration
       );
     }
+  }
+
+  /**
+   * Awaits for _this_ Map instance to be "loaded" and returns a Promise to the Map.
+   * If _this_ Map instance is already loaded, the Promise is resolved directly,
+   * otherwise, it is resolved as a result of the "load" event.
+   * @returns
+   */
+  async onLoadAsync() {
+    return new Promise<Map>((resolve, reject) => {
+      if (this.loaded()) {
+        return resolve(this);
+      }
+
+      this.once("load", (_) => {
+        resolve(this);
+      });
+    });
+  }
+
+  /**
+   * Awaits for _this_ Map instance to be "loaded" as well as with terrain being non-null for the first time
+   * and returns a Promise to the Map.
+   * If _this_ Map instance is already loaded with terrain, the Promise is resolved directly,
+   * otherwise, it is resolved as a result of the "loadWithTerrain" event.
+   * @returns
+   */
+  async onLoadWithTerrainAsync() {
+    return new Promise<Map>((resolve, reject) => {
+      if (this.loaded() && this.terrain) {
+        return resolve(this);
+      }
+
+      this.once("loadWithTerrain", (_) => {
+        resolve(this);
+      });
+    });
   }
 
   /**
@@ -846,6 +943,7 @@ export class Map extends maplibregl.Map {
         this.terrainFlattening = false;
         this.terrain.exaggeration = exaggeration;
       }
+
       this.triggerRepaint();
     };